refactor: 유틸 파일 구조 변경
